# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NBA SIM API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the NBA SIM PROJECT."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/nba",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/nba/match/play/{id}": {
      "post": {
        "summary": "Simulate and play a match. (Coaches & Admin)",
        "tags": [
          "Matches",
          "Coaches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Match simulated and sent to Commissioner for approval.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Match simulated."
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/nba/player/": {
      "post": {
        "summary": "Create a new player (Coach or Admin only)",
        "description": "Creates a new player and automatically sends the player record to the Commissioner (Admin) for approval.   Requires authentication and one of the roles: **coach**, **admin**.\n",
        "tags": [
          "Coaches",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "currentTeam",
                  "position",
                  "possession",
                  "three",
                  "layup",
                  "defense"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "currentTeam": {
                    "type": "string"
                  },
                  "position": {
                    "type": "string"
                  },
                  "possession": {
                    "type": "number"
                  },
                  "three": {
                    "type": "number"
                  },
                  "layup": {
                    "type": "number"
                  },
                  "defense": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Player created and sent for Commissioner approval."
          }
        }
      },
      "get": {
        "summary": "Retrieve all approved active players (Users + Admin)",
        "description": "Returns all **approved and active** players.   Accessible to roles: **user**, **admin**.\n",
        "tags": [
          "Users",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Approved active players returned."
          }
        }
      }
    },
    "/api/nba/player/admin/": {
      "get": {
        "summary": "Retrieve all players (approved + pending) — Admin only",
        "description": "Returns all players in the system regardless of review status.   Restricted to **admin** users.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All players retrieved successfully."
          }
        }
      }
    },
    "/api/nba/player/pending/": {
      "get": {
        "summary": "Retrieve all pending players — Admin only",
        "description": "Retrieves players still awaiting Commissioner review.   Restricted to **admin** users only.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pending players returned."
          }
        }
      }
    },
    "/api/nba/player/review/{id}": {
      "put": {
        "summary": "Approve or reject a pending player — Admin only",
        "description": "Allows the Commissioner (Admin) to approve or reject a pending player.   When approved, the player becomes active.   When rejected, the player is returned to the coach.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player ID to review"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "approved"
                ],
                "properties": {
                  "approved": {
                    "type": "boolean",
                    "description": "Approve (true) or reject (false) the player"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player review status updated."
          }
        }
      }
    },
    "/api/nba/player/update/{id}": {
      "put": {
        "summary": "Update an existing player (Coach or Admin)",
        "description": "Updates a player and automatically deactivates them, sending the updated record for Commissioner review.   Accessible to roles **coach** and **admin**.\n",
        "tags": [
          "Coaches",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the player to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "position": {
                    "type": "string"
                  },
                  "currentTeam": {
                    "type": "string"
                  },
                  "possession": {
                    "type": "number"
                  },
                  "three": {
                    "type": "number"
                  },
                  "layup": {
                    "type": "number"
                  },
                  "defense": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player updated and sent for re-approval."
          }
        }
      }
    },
    "/api/nba/player/{id}": {
      "get": {
        "summary": "Get a player by ID (User + Admin)",
        "description": "Retrieves a single approved player by ID.   Accessible to **user** and **admin** roles.\n",
        "tags": [
          "Users",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the player"
          }
        ],
        "responses": {
          "200": {
            "description": "Player found and returned."
          }
        }
      }
    },
...
    "/api/nba/team/player/{id}": {
      "put": {
        "summary": "Add or update a player's team assignment (Coach or Admin)",
        "description": "Adds a player to a team. The player is deactivated and sent to the Commissioner for review.",
        "tags": [
          "Teams",
          "Coaches",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "playerId"
                ],
                "properties": {
                  "playerId": {
                    "type": "string",
                    "example": "player789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player assigned or updated for the team."
          }
        }
      },
      "delete": {
        "summary": "Remove a player from a team (Coach or Admin)",
        "description": "Removes a player from a team. The player is deactivated and sent to the Commissioner for review.",
        "tags": [
          "Teams",
          "Coaches",
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "playerId"
                ],
                "properties": {
                  "playerId": {
                    "type": "string",
                    "example": "player789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player removed and record updated."
          }
        }
      }
    },
...
  },
...
}